+++ Frends.JSON.Transform/Frends.JSON.Transform.Tests/UnitTests.cs
@@ -1,5 +1,6 @@
 using NUnit.Framework;
 using System;
+using System.Threading;
 using Newtonsoft.Json.Linq;
 using Frends.JSON.Transform.Definitions;
 
@@ -40,21 +41,21 @@ class TestClass
     public void TransformShouldAllowJTokenAsInput()
     {
         _testInput.InputJson = JToken.Parse(_testJson);
-        var result = JSON.Transform(_testInput);
+        var result = JSON.Transform(_testInput, new Options(), new CancellationToken());
         Assert.AreEqual("{\"FullName\":\"Veijo Frends\",\"Age\":30,\"StillBreething\":false}", result.Transformation);
     }
 
     [Test]
     public void TransformShouldAllowStringAsInput()
     {
-        var result = JSON.Transform(_testInput);
+        var result = JSON.Transform(_testInput, new Options(), new CancellationToken());
         Assert.AreEqual("{\"FullName\":\"Veijo Frends\",\"Age\":30,\"StillBreething\":false}", result.Transformation);
     }
 
     [Test]
     public void TransformMapsStringData()
     {
-        var result = JSON.Transform(_testInput);
+        var result = JSON.Transform(_testInput, new Options(), new CancellationToken());
 
         var fullName = result.JToken.FullName;
 
@@ -64,7 +65,7 @@ class TestClass
     [Test]
     public void TransformMapsNumbersCorrectly()
     {
-        var result = JSON.Transform(_testInput);
+        var result = JSON.Transform(_testInput, new Options(), new CancellationToken());
 
         var age = result.JToken.Age;
 
@@ -75,7 +76,7 @@ class TestClass
     [Test]
     public void TransformationMapsBoolValueCorrectly()
     {
-        var result = JSON.Transform(_testInput);
+        var result = JSON.Transform(_testInput, new Options(), new CancellationToken());
 
         var breething = result.JToken.StillBreething;
 
@@ -89,7 +90,7 @@ class TestClass
         _testInput.InputJson = @"{ ""array"": [{""key"":""first element""},{""key"":""second element""}]}";
         _testInput.JsonMap = @"{""firstElement"":""#valueof($.array[0].key)""}";
 
-        var result = JSON.Transform(_testInput);
+        var result = JSON.Transform(_testInput, new Options(), new CancellationToken());
         var firstElement = result.JToken.firstElement;
 
         Assert.AreEqual("first element", (string)firstElement);
@@ -102,7 +103,7 @@ class TestClass
         _testInput.InputJson = @"[{""key"": ""first element""}, {""key"": ""second element""}]";
         _testInput.JsonMap = @"{""firstElement"": ""#valueof($.[0].key)""}";
 
-        var ex = Assert.Throws<Exception>(() => JSON.Transform(_testInput));
+        var ex = Assert.Throws<Exception>(() => JSON.Transform(_testInput, new Options(), new CancellationToken()));
         Assert.That(ex.Message.Equals("Json transformation failed: Input Json is not valid: Array is not supported as root element."));
     }
 
@@ -112,7 +113,7 @@ class TestClass
     {
         _testInput.InputJson = @"{ foo baar";
 
-        Assert.Throws<FormatException>(() => JSON.Transform(_testInput));
+        Assert.Throws<FormatException>(() => JSON.Transform(_testInput, new Options(), new CancellationToken()));
     }
 
     [Test]
@@ -120,7 +121,7 @@ class TestClass
     {
         _testInput.JsonMap = @"{""age"":""#valuof($.age)"", ""foo}";
 
-        Assert.Throws<Exception>(() => JSON.Transform(_testInput));
+        Assert.Throws<Exception>(() => JSON.Transform(_testInput, new Options(), new CancellationToken()));
     }
 }